+++ contrib/platform/test/com/sun/jna/platform/win32/Ole32Test.java
@@ -16,7 +16,9 @@ import junit.framework.TestCase;
 
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
+import com.sun.jna.platform.win32.BaseTSD.SIZE_T;
 import com.sun.jna.platform.win32.Guid.GUID;
+import com.sun.jna.platform.win32.WinDef.LPVOID;
 import com.sun.jna.platform.win32.WinNT.HRESULT;
 import com.sun.jna.ptr.PointerByReference;
 
@@ -106,4 +108,24 @@ public class Ole32Test extends TestCase {
 	    assertEquals(WinError.S_OK, Ole32.INSTANCE.CLSIDFromProgID("jpegfile", clsid));
 	    assertEquals("{25336920-03F9-11CF-8FD0-00AA00686F13}", clsid.toGuidString());
 	}
+
+    public void testCoTaskMemAlloc() {
+        LPVOID ptr = Ole32.INSTANCE.CoTaskMemAlloc(new SIZE_T(256));
+
+        assertTrue(ptr.longValue() != 0);
+
+        Ole32.INSTANCE.CoTaskMemFree(ptr);
+    }
+
+    public void testCoTaskMemRealloc() {
+        LPVOID ptr = Ole32.INSTANCE.CoTaskMemAlloc(new SIZE_T(256));
+
+        assertTrue(ptr.longValue() != 0);
+
+        ptr = Ole32.INSTANCE.CoTaskMemRealloc(ptr, new SIZE_T(128));
+
+        assertTrue(ptr.longValue() != 0);
+
+        Ole32.INSTANCE.CoTaskMemFree(ptr);
+    }
 }